refactor: move bili-api to external modules
