+++ package.json
@@ -30,6 +30,7 @@
     "@typescript-eslint/parser": "^4.23.0",
     "assert-never": "^1.2.1",
     "axios": "^0.21.1",
+    "bili-api": "github:AidenLx/bili-api",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^7.26.0",
     "eslint-config-prettier": "^8.3.0",
